Walmart Sales Dataset of 45 stores
In this project, I tried to answer 5 questions that were asked about the data which are:
Q1: Which store has the maximum sales in this dataset?
Q2: Which store has the maximum standard deviation?
Q3: Some holidays have a negative impact on sales. Find out holidays that have higher sales than the mean sales in the non-holiday season for all stores together
Q4: Provide a monthly and semester view of sales in units and give insights
Q5: Plot the relations between weekly sales vs. other numeric features and give insights.